GEODE-8784:Create matching Geode stats for all Redis INFO stats - move to statistics package
